Allow saving image attachments to album

+++ src/main/java/smithereen/storage/media/LocalMediaFileStorageDriver.java
@@ -14,19 +14,21 @@ import smithereen.Config;
 import smithereen.Utils;
 import smithereen.model.ObfuscatedObjectIDType;
 import smithereen.model.media.MediaFileID;
-import smithereen.model.media.MediaFileRecord;
 import smithereen.storage.ImgProxy;
 import smithereen.util.XTEA;
 
 public class LocalMediaFileStorageDriver extends MediaFileStorageDriver{
 	@Override
-	public void storeFile(File localFile, MediaFileID id) throws IOException{
+	public void storeFile(File localFile, MediaFileID id, boolean keepLocalFile) throws IOException{
 		int oid=Math.abs(id.originalOwnerID());
 		File dir=new File(Config.uploadPath, String.format(Locale.US, "%02d/%02d/%02d", oid%100, oid/100%100, oid/10000%100));
 		if(!dir.exists() && !dir.mkdirs())
 			throw new IOException("Failed to create directories");
 		File targetFile=new File(Config.uploadPath, getFilePath(id));
-		Files.move(localFile.toPath(), targetFile.toPath());
+		if(keepLocalFile)
+			Files.copy(localFile.toPath(), targetFile.toPath());
+		else
+			Files.move(localFile.toPath(), targetFile.toPath());
 	}
 
 	@Override
